fix: move logs out of inngest function
