+++ .env.example
@@ -129,6 +129,10 @@ TWILIO_ACCOUNT_SID=
 TWILIO_AUTH_TOKEN=
 TWILIO_PHONE_NUMBER=
 
+# Custom Stop/Unstop keyowrd used in the SMS flow
+SWC_STOP_SMS_KEYWORD=
+SWC_UNSTOP_SMS_KEYWORD=
+
 # Engine used to airdrop NFTs
 THIRDWEB_ENGINE_URL=
 THIRDWEB_ENGINE_ACCESS_TOKEN=
