+++ .vscode/settings.json
@@ -67,6 +67,7 @@
     "STOPALL",
     "subpremise",
     "thirdweb",
+    "twiml",
     "typecheck",
     "undici",
     "unoptimized",
